Preparing for translation Java -> Kotlin
