+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".JavaActivity"/>
+        <activity android:name=".KotlinActivity"/>
     </application>
 
 </manifest>
