+++ app/src/main/java/com/android/androidonkotlin/MainActivity.java
@@ -2,6 +2,7 @@ package com.android.androidonkotlin;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
@@ -9,35 +10,24 @@ import android.widget.TextView;
 
 public class MainActivity extends AppCompatActivity {
 
-    private Button minusButton = null;
-    private Button plusButton = null;
-    private TextView textViewCounter = null;
-
-    private int counter = 0;
+    private Button javaButton = null;
+    private Button kotlinButton = null;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate ( savedInstanceState );
-        setContentView ( R.layout.activity_main );
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
 
-        minusButton = findViewById ( R.id.button_minus );
-        plusButton = findViewById ( R.id.button_plus );
-        textViewCounter = findViewById ( R.id.text_view_counter );
+        javaButton = findViewById(R.id.button_java);
+        javaButton.setOnClickListener(view -> {
+            Intent intent = new Intent(MainActivity.this, JavaActivity.class);
+            startActivity(intent);
+        });
 
-        counterUpdate(counter);
-
-        minusButton.setOnClickListener ( new View.OnClickListener () {
-            @Override
-            public void onClick(View v) {
-                counterUpdate ( --counter );
-            }
-        } );
-
-        plusButton.setOnClickListener ( v ->
-                counterUpdate ( ++counter ) );
-    }
-
-    private  void counterUpdate (int counter){
-        textViewCounter.setText ( String.valueOf ( counter ) );
+        kotlinButton = findViewById(R.id.button_kotlin);
+        kotlinButton.setOnClickListener(view -> {
+            Intent intent = new Intent(MainActivity.this, KotlinActivity.class);
+            startActivity(intent);
+        });
     }
 }
