+++ app/src/main/res/layout/activity_counter.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="horizontal"
+    tools:context=".MainActivity">
+
+    <Button
+        android:id="@+id/button_minus"
+        style="@style/button_counter"
+        android:text="@string/minus" />
+
+    <TextView
+        android:id="@+id/text_view_counter"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:layout_weight="2"
+        android:gravity="center"
+        android:text="@string/text_view"
+        android:textSize="@dimen/text_view_counter_main"
+        android:textStyle="bold"
+        tools:ignore="SpUsage" />
+
+    <Button
+        android:id="@+id/button_plus"
+        style="@style/button_counter"
+        android:text="@string/plus" />
+
+</LinearLayout>
