+++ app/src/main/res/layout/activity_main.xml
@@ -1,32 +1,28 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="horizontal"
+    android:orientation="vertical"
     tools:context=".MainActivity">
 
     <Button
-        android:id="@+id/button_minus"
-        style="@style/button_main"
-        android:text="@string/minus" />
-
-    <TextView
-        android:id="@+id/text_view_counter"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_weight="2"
-        android:gravity="center"
-        android:text="@string/text_view"
-        android:textSize="@dimen/text_view_counter_main"
-        android:textStyle="bold"
-        tools:ignore="SpUsage" />
+        android:id="@+id/button_java"
+        app:layout_constraintBottom_toTopOf="@id/button_kotlin"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        android:text="@string/java"
+        style="@style/button_main" />
 
     <Button
-        android:id="@+id/button_plus"
+        android:id="@+id/button_kotlin"
         style="@style/button_main"
-        android:text="@string/plus" />
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/button_java"
+        android:text="@string/kotlin" />
 
-</LinearLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
