+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,6 @@
     <string name="minus">-\nминус</string>
     <string name="plus">+ \n плюс</string>
     <string name="text_view">0</string>
+    <string name="java">Java</string>
+    <string name="kotlin">Kotlin</string>
 </resources>
